Update
Fix some parts of the user backend
Complete Profile Modal now uses /users/profile fetch to check if the user has complete profile, if not a pop up shows up.
Business side is yet to be connected to the db.
My-pets has problem with the backend, but frontend is fine
user/pet Profile pic is yet to work on the backend.