Update
Lots of major update regarding app behavior in the mobile environment and added chat function.
Please test the functionality of the each added updates. because a conflict arouse wherein the backend was not updated during the past git push. after this git is push was the update of the backend was performed.
Take note of the bugs and problem that will be encountered when testing and compile the result.