Update
Added back buttons on index pages in which retuns to the home page by default.

revise the payment-gateway, added payment preference and change the payment method to a dropdown.

Revise ratings review design.

Adjusted verification-documents to handle 3 document permits based on the web.

Re-align operating hours in quick action with the quick action in the business info.

revise tags in profile, made it bento layout (di mukhang bento haha) and with different colors.

Adjust the Business description to have an accordion for lengthy description.

daming changes  but so far all end well.