Update
- Greatly revamped both business/user side AppointmentDetails. (on-going).
- Greatly revamped business side booking index (user side next).
- Styled alerts/pop up modals on AppointmentDetails, booking index, Schedule Details and upload-payment-proof.

To-do-Next:
- Check the Approve/Reject alert/pop-up modal and style it.
- Push-up notification should have the main features per user sides:

User side: push up notif:
When business leaves a message.
When their appointment status changes.
When their resched status changes.
When their payment status changes.
When their appointment is getting close
When they payment is still not due

Business side:
When a user books an appointment
When a user sends a message
When a user uploads payment
When a user rescheds
When a user leaves a rating
When a booked service is drawing close.

To-do-Next:
Add a share feature wherein the user can share the service/business outside the app. like in messenger, email etc.,
Check the alert/modal when the approve/reject button is pressed on the resched appointment.
Continue to simulate the system flow and redesign the layout and styles as necessary.
Implement the Push up notification.